requirements.txt
fuzzywuzzy==0.18.0
discord.py==2.1.0
aiohttp==3.8.5
python-Levenshtein==0.12.2
datetime==5.4
